embedfs: Optimize a little to improve performance and add mutex where needed
lock: Replace `lock_dec_spinning()` with `mutex_lock_atomic()`

buffer-cache: Add a variable, size, in `buffer_cache` structure

  This is preparation to support variable size buffer for device.
  Especially when it comes to small ram and relatively bigger flash
  block, it consumes too much memory allocating buffer-cache for flash
  block. Make buffer-cache works even when it has smaller size than a
  flash block.

stm32: Separate semaphore implementation in a file, no more in `lock.h`
as inline function

Polish waitqueue

+++ lib/memcpy.c
@@ -1,4 +1,6 @@
-void *memcpy(void *dst, const void *src, int len)
+#include <types.h>
+
+void *memcpy(void *dst, const void *src, size_t len)
 {
 	const char *s = src;
 	char *d = dst;
